refactor: Clean code apply + CategoryModel ID and TaskModel ID replaced by Integer type

+++ app/src/main/java/com/example/tasks/adapter/FinishedTaskAdapter.java
@@ -52,7 +52,7 @@ public class FinishedTaskAdapter extends RecyclerView.Adapter<FinishedTaskAdapte
             int catAdaptPosition,
             ActivityResultLauncher<Intent> actResult,
             SQLiteHelper myDB,
-            long categoryId
+            Integer categoryId
     ) {
         this.activity = activity;
         this.catAdaptPosition = catAdaptPosition;
